chore: add packages, restructure classes
