+++ src/main/java/io/camunda/connector/slack/dto/Conversation.java
@@ -1,0 +1,50 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. Licensed under a proprietary license.
+ * See the License.txt file for more information. You may not use this file
+ * except in compliance with the proprietary license.
+ */
+package io.camunda.connector.slack.dto;
+
+import java.util.Objects;
+
+public class Conversation {
+
+  private final String id;
+  private final String name;
+
+  public Conversation(com.slack.api.model.Conversation conversation) {
+    this.id = conversation.getId();
+    this.name = conversation.getName();
+  }
+
+  public String getId() {
+    return id;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    Conversation that = (Conversation) o;
+    return Objects.equals(id, that.id) && Objects.equals(name, that.name);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(id, name);
+  }
+
+  @Override
+  public String toString() {
+    return "Conversation{" + "id='" + id + '\'' + ", name='" + name + '\'' + '}';
+  }
+}
