+++ src/main/java/io/camunda/connector/slack/model/ConversationsCreateSlackResponse.java
@@ -4,9 +4,11 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.slack;
+package io.camunda.connector.slack.model;
 
 import com.slack.api.methods.response.conversations.ConversationsCreateResponse;
+import io.camunda.connector.slack.SlackResponse;
+import io.camunda.connector.slack.dto.Conversation;
 import java.util.Objects;
 
 public class ConversationsCreateSlackResponse implements SlackResponse {
