+++ src/test/java/io/camunda/connector/slack/BaseTest.java
@@ -11,6 +11,7 @@ import static java.nio.file.Files.readString;
 
 import com.google.gson.Gson;
 import io.camunda.connector.api.outbound.OutboundConnectorContext;
+import io.camunda.connector.slack.suppliers.GsonSupplier;
 import io.camunda.connector.test.outbound.OutboundConnectorContextBuilder;
 import java.io.File;
 import java.io.IOException;
