+++ src/test/java/io/camunda/connector/slack/SlackFunctionTest.java
@@ -27,6 +27,9 @@ import com.slack.api.model.Message;
 import com.slack.api.model.ResponseMetadata;
 import com.slack.api.model.User;
 import io.camunda.connector.api.outbound.OutboundConnectorContext;
+import io.camunda.connector.slack.model.ChatPostMessageSlackResponse;
+import io.camunda.connector.slack.model.ConversationsCreateData;
+import io.camunda.connector.slack.model.ConversationsCreateSlackResponse;
 import java.io.IOException;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
